feat: refactor update source option types and hook for improved structure and clarity
